+++ lib/model.d.ts
@@ -8,7 +8,6 @@ import {DataType} from './data-types';
 import {Sequelize} from './sequelize';
 import {AbstractDeferrable} from './deferrable';
 import {ModelManager} from './model-manager';
-import {Logging, Transactionable, SearchPathable, Filterable, Projectable, Paranoid} from './misc-types';
 import {
   Association,
   BelongsTo,
@@ -20,6 +19,53 @@ import {
   BelongsToMany,
   BelongsToManyOptions
 } from './associations/index';
+
+export interface Logging {
+  /**
+   * A function that gets executed while running the query to log the sql.
+   */
+  logging?: boolean | Function;
+}
+
+export interface Transactionable {
+  /**
+   * Transaction to run query under
+   */
+  transaction?: Transaction;
+}
+
+export interface SearchPathable {
+  /**
+   * An optional parameter to specify the schema search_path (Postgres only)
+   */
+  searchPath?: string;
+}
+
+export interface Filterable {
+  /**
+   * Attribute has to be matched for rows to be selected for the given action.
+   */
+  where?: WhereOptions;
+}
+
+export interface Projectable {
+  /**
+   * A list of the attributes that you want to select. To rename an attribute, you can pass an array, with
+   * two elements - the first is the name of the attribute in the DB (or some kind of expression such as
+   * `Sequelize.literal`, `Sequelize.fn` and so on), and the second is the name you want the attribute to
+   * have in the returned instance
+   */
+  attributes?: FindAttributeOptions;
+}
+
+export interface Paranoid {
+    /**
+   * If true, only non-deleted records will be returned. If false, both deleted and non-deleted records will
+   * be returned. Only applies if `options.paranoid` is true for the model.
+   */
+  paranoid?: boolean;
+}
+
 
 export type GroupOption = string | Fn | Col | (string | Fn | Col)[];
